+++ packages/storage/src/bucket.rs
@@ -14,21 +14,21 @@ use crate::type_helpers::deserialize_kv;
 use crate::type_helpers::{may_deserialize, must_deserialize};
 
 /// An alias of Bucket::new for less verbose usage
-pub fn bucket<'a, S, T>(namespace: &[u8], storage: &'a mut S) -> Bucket<'a, S, T>
+pub fn bucket<'a, S, T>(storage: &'a mut S, namespace: &[u8]) -> Bucket<'a, S, T>
 where
     S: Storage,
     T: Serialize + DeserializeOwned,
 {
-    Bucket::new(namespace, storage)
+    Bucket::new(storage, namespace)
 }
 
 /// An alias of ReadonlyBucket::new for less verbose usage
-pub fn bucket_read<'a, S, T>(namespace: &[u8], storage: &'a S) -> ReadonlyBucket<'a, S, T>
+pub fn bucket_read<'a, S, T>(storage: &'a S, namespace: &[u8]) -> ReadonlyBucket<'a, S, T>
 where
     S: ReadonlyStorage,
     T: Serialize + DeserializeOwned,
 {
-    ReadonlyBucket::new(namespace, storage)
+    ReadonlyBucket::new(storage, namespace)
 }
 
 pub struct Bucket<'a, S, T>
@@ -47,7 +47,7 @@ where
     S: Storage,
     T: Serialize + DeserializeOwned,
 {
-    pub fn new(namespace: &[u8], storage: &'a mut S) -> Self {
+    pub fn new(storage: &'a mut S, namespace: &[u8]) -> Self {
         Bucket {
             storage,
             prefix: to_length_prefixed(namespace),
@@ -55,7 +55,7 @@ where
         }
     }
 
-    pub fn multilevel(namespaces: &[&[u8]], storage: &'a mut S) -> Self {
+    pub fn multilevel(storage: &'a mut S, namespaces: &[&[u8]]) -> Self {
         Bucket {
             storage,
             prefix: to_length_prefixed_nested(namespaces),
@@ -130,7 +130,7 @@ where
     S: ReadonlyStorage,
     T: Serialize + DeserializeOwned,
 {
-    pub fn new(namespace: &[u8], storage: &'a S) -> Self {
+    pub fn new(storage: &'a S, namespace: &[u8]) -> Self {
         ReadonlyBucket {
             storage,
             prefix: to_length_prefixed(namespace),
@@ -138,7 +138,7 @@ where
         }
     }
 
-    pub fn multilevel(namespaces: &[&[u8]], storage: &'a S) -> Self {
+    pub fn multilevel(storage: &'a S, namespaces: &[&[u8]]) -> Self {
         ReadonlyBucket {
             storage,
             prefix: to_length_prefixed_nested(namespaces),
@@ -188,7 +188,7 @@ mod test {
     #[test]
     fn store_and_load() {
         let mut store = MockStorage::new();
-        let mut bucket = bucket::<_, Data>(b"data", &mut store);
+        let mut bucket = bucket::<_, Data>(&mut store, b"data");
 
         // save data
         let data = Data {
@@ -205,7 +205,7 @@ mod test {
     #[test]
     fn remove_works() {
         let mut store = MockStorage::new();
-        let mut bucket = bucket::<_, Data>(b"data", &mut store);
+        let mut bucket = bucket::<_, Data>(&mut store, b"data");
 
         // save data
         let data = Data {
@@ -227,7 +227,7 @@ mod test {
     #[test]
     fn readonly_works() {
         let mut store = MockStorage::new();
-        let mut bucket = bucket::<_, Data>(b"data", &mut store);
+        let mut bucket = bucket::<_, Data>(&mut store, b"data");
 
         // save data
         let data = Data {
@@ -236,7 +236,7 @@ mod test {
         };
         bucket.save(b"maria", &data).unwrap();
 
-        let reader = bucket_read::<_, Data>(b"data", &mut store);
+        let reader = bucket_read::<_, Data>(&mut store, b"data");
 
         // check empty data handling
         assert!(reader.load(b"john").is_err());
@@ -250,7 +250,7 @@ mod test {
     #[test]
     fn buckets_isolated() {
         let mut store = MockStorage::new();
-        let mut bucket1 = bucket::<_, Data>(b"data", &mut store);
+        let mut bucket1 = bucket::<_, Data>(&mut store, b"data");
 
         // save data
         let data = Data {
@@ -259,7 +259,7 @@ mod test {
         };
         bucket1.save(b"maria", &data).unwrap();
 
-        let mut bucket2 = bucket::<_, Data>(b"dat", &mut store);
+        let mut bucket2 = bucket::<_, Data>(&mut store, b"dat");
 
         // save data (dat, amaria) vs (data, maria)
         let data2 = Data {
@@ -269,14 +269,14 @@ mod test {
         bucket2.save(b"amaria", &data2).unwrap();
 
         // load one
-        let reader = bucket_read::<_, Data>(b"data", &store);
+        let reader = bucket_read::<_, Data>(&store, b"data");
         let loaded = reader.load(b"maria").unwrap();
         assert_eq!(data, loaded);
         // no cross load
         assert_eq!(None, reader.may_load(b"amaria").unwrap());
 
         // load the other
-        let reader2 = bucket_read::<_, Data>(b"dat", &store);
+        let reader2 = bucket_read::<_, Data>(&store, b"dat");
         let loaded2 = reader2.load(b"amaria").unwrap();
         assert_eq!(data2, loaded2);
         // no cross load
@@ -286,7 +286,7 @@ mod test {
     #[test]
     fn update_success() {
         let mut store = MockStorage::new();
-        let mut bucket = bucket::<_, Data>(b"data", &mut store);
+        let mut bucket = bucket::<_, Data>(&mut store, b"data");
 
         // initial data
         let init = Data {
@@ -316,7 +316,7 @@ mod test {
     #[test]
     fn update_can_change_variable_from_outer_scope() {
         let mut store = MockStorage::new();
-        let mut bucket = bucket::<_, Data>(b"data", &mut store);
+        let mut bucket = bucket::<_, Data>(&mut store, b"data");
         let init = Data {
             name: "Maria".to_string(),
             age: 42,
@@ -339,7 +339,7 @@ mod test {
     #[test]
     fn update_fails_on_error() {
         let mut store = MockStorage::new();
-        let mut bucket = bucket::<_, Data>(b"data", &mut store);
+        let mut bucket = bucket::<_, Data>(&mut store, b"data");
 
         // initial data
         let init = Data {
@@ -374,7 +374,7 @@ mod test {
         }
 
         let mut store = MockStorage::new();
-        let mut bucket = bucket::<_, Data>(b"data", &mut store);
+        let mut bucket = bucket::<_, Data>(&mut store, b"data");
 
         // initial data
         let init = Data {
@@ -408,7 +408,7 @@ mod test {
     #[test]
     fn update_handles_on_no_data() {
         let mut store = MockStorage::new();
-        let mut bucket = bucket::<_, Data>(b"data", &mut store);
+        let mut bucket = bucket::<_, Data>(&mut store, b"data");
 
         let init_value = Data {
             name: "Maria".to_string(),
@@ -433,7 +433,7 @@ mod test {
     #[cfg(feature = "iterator")]
     fn range_over_data() {
         let mut store = MockStorage::new();
-        let mut bucket = bucket::<_, Data>(b"data", &mut store);
+        let mut bucket = bucket::<_, Data>(&mut store, b"data");
 
         let jose = Data {
             name: "Jose".to_string(),
@@ -455,7 +455,7 @@ mod test {
         assert_eq!(data[1], (b"maria".to_vec(), maria.clone()));
 
         // also works for readonly
-        let read_bucket = bucket_read::<_, Data>(b"data", &store);
+        let read_bucket = bucket_read::<_, Data>(&store, b"data");
         let res_data: StdResult<Vec<KV<Data>>> =
             read_bucket.range(None, None, Order::Ascending).collect();
         let data = res_data.unwrap();